# internal/api/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Asynchronous DNS testing with support for Do53, DoT, DoH, DoQ protocols\nThis API allows you to submit DNS queries and retrieve results asynchronously",
        "title": "DNS-Tester-GO API",
        "contact": {
            "name": "DNS-Tester-GO",
            "url": "https://github.com/sudo-Tiz/DNS-Tester-GO",
            "email": "contact@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://github.com/sudo-Tiz/DNS-Tester-GO/blob/main/LICENSE"
        },
        "version": "1.0.0"
    },
    "host": "localhost:5000",
    "basePath": "/",
    "paths": {
        "/dns-lookup": {
            "post": {
                "description": "Enqueue a DNS lookup for asynchronous processing. Returns a task ID that can be polled.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DNS"
                ],
                "summary": "Submit DNS lookup task",
                "parameters": [
                    {
                        "description": "DNS lookup parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.DNSLookupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task accepted and enqueued",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.TaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request or missing parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "No workers available",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Check if the API service is running and workers are available",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy or degraded",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/metrics": {
            "get": {
                "description": "Expose application metrics in Prometheus format",
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Prometheus metrics",
                "responses": {
                    "200": {
                        "description": "Prometheus metrics",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/reverse-lookup": {
            "post": {
                "description": "Enqueue a reverse DNS lookup for an IP address. Automatically converts IP to PTR format.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "DNS"
                ],
                "summary": "Submit reverse DNS lookup (PTR)",
                "parameters": [
                    {
                        "description": "Reverse lookup parameters",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ReverseLookupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task accepted and enqueued",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.TaskResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid IP address or missing parameters",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Rate limit exceeded",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "No workers available",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/status": {
            "get": {
                "description": "Check if the API service is running and workers are available",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "System"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "Service is healthy or degraded",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.HealthResponse"
                        }
                    }
                }
            }
        },
        "/tasks/{taskID}": {
            "get": {
                "description": "Retrieve the status and result of a previously submitted DNS lookup task",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task status and result",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Task ID",
                        "name": "taskID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Task found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.TaskStatusResponse"
                        }
                    },
                    "404": {
                        "description": "Task not found",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_sudo-tiz_dns-tester-go_internal_models.DNSAnswer": {
            "description": "DNS resource record with name, type, TTL, and value",
            "type": "object",
            "properties": {
                "name": {
                    "description": "DNS name",
                    "type": "string",
                    "example": "example.com."
                },
                "ttl": {
                    "description": "Time to live in seconds",
                    "type": "integer",
                    "example": 3600
                },
                "type": {
                    "description": "Record type",
                    "type": "string",
                    "example": "A"
                },
                "value": {
                    "description": "Record value",
                    "type": "string",
                    "example": "93.184.216.34"
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.DNSLookupRequest": {
            "description": "DNS lookup request with domain, query type, and optional DNS servers",
            "type": "object",
            "required": [
                "domain",
                "qtype"
            ],
            "properties": {
                "dns_servers": {
                    "description": "DNS servers to query (optional, uses config if empty)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.DNSServer"
                    }
                },
                "domain": {
                    "description": "Domain name to query",
                    "type": "string",
                    "example": "example.com"
                },
                "qtype": {
                    "description": "Query type (A, AAAA, MX, TXT, etc.)",
                    "type": "string",
                    "example": "A"
                },
                "tls_insecure_skip_verify": {
                    "description": "Skip TLS certificate verification (testing only)",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.DNSLookupResult": {
            "description": "Result from a single DNS server query",
            "type": "object",
            "properties": {
                "answers": {
                    "description": "DNS answers",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.DNSAnswer"
                    }
                },
                "command_status": {
                    "description": "Command execution status",
                    "type": "string",
                    "example": "success"
                },
                "dns_protocol": {
                    "description": "Protocol used (udp, tcp, tls, https, quic)",
                    "type": "string",
                    "example": "udp"
                },
                "error": {
                    "description": "Error message if query failed",
                    "type": "string",
                    "example": "connection timeout"
                },
                "name": {
                    "description": "Queried name",
                    "type": "string",
                    "example": "example.com."
                },
                "qtype": {
                    "description": "Query type",
                    "type": "string",
                    "example": "A"
                },
                "rcode": {
                    "description": "DNS response code",
                    "type": "string",
                    "example": "NOERROR"
                },
                "tags": {
                    "description": "Server tags",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "GOOGLE",
                        "PRIMARY"
                    ]
                },
                "time_ms": {
                    "description": "Query execution time in milliseconds",
                    "type": "number",
                    "example": 23.45
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.DNSLookupResults": {
            "description": "Aggregated DNS lookup results from all queried servers",
            "type": "object",
            "properties": {
                "details": {
                    "description": "Results per DNS server (keyed by target)",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.DNSLookupResult"
                    }
                },
                "duration": {
                    "description": "Total query duration in seconds",
                    "type": "number",
                    "example": 0.125
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.DNSServer": {
            "description": "DNS server configuration with protocol://host:port format",
            "type": "object",
            "properties": {
                "tags": {
                    "description": "Optional tags for identification",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "GOOGLE",
                        "PRIMARY",
                        "PUBLIC"
                    ]
                },
                "target": {
                    "description": "DNS server in format protocol://host:port",
                    "type": "string",
                    "example": "udp://8.8.8.8:53"
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.ErrorResponse": {
            "description": "Error response returned for failed requests",
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error message",
                    "type": "string",
                    "example": "rate limit exceeded"
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.HealthResponse": {
            "description": "Health check response",
            "type": "object",
            "properties": {
                "status": {
                    "description": "Health status (ok, degraded)",
                    "type": "string",
                    "example": "ok"
                },
                "warning": {
                    "description": "Warning message if degraded",
                    "type": "string",
                    "example": "no active workers detected"
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.ReverseLookupRequest": {
            "description": "Reverse DNS lookup request for an IP address",
            "type": "object",
            "required": [
                "reverse_ip"
            ],
            "properties": {
                "dns_servers": {
                    "description": "DNS servers to query (optional)",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.DNSServer"
                    }
                },
                "reverse_ip": {
                    "description": "IP address to reverse lookup",
                    "type": "string",
                    "example": "8.8.8.8"
                },
                "tls_insecure_skip_verify": {
                    "description": "Skip TLS certificate verification",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.TaskResponse": {
            "description": "Task submission response with unique task ID",
            "type": "object",
            "properties": {
                "message": {
                    "description": "Status message",
                    "type": "string",
                    "example": "DNS lookup enqueued"
                },
                "task_id": {
                    "description": "Unique task identifier for polling",
                    "type": "string",
                    "example": "abc123def456789"
                }
            }
        },
        "github_com_sudo-tiz_dns-tester-go_internal_models.TaskStatusResponse": {
            "description": "Task status response with result when completed",
            "type": "object",
            "properties": {
                "completed_at": {
                    "description": "Task completion timestamp",
                    "type": "string"
                },
                "created_at": {
                    "description": "Task creation timestamp",
                    "type": "string"
                },
                "error": {
                    "description": "Error message (populated when status is FAILURE)",
                    "type": "string",
                    "example": "worker timeout"
                },
                "task_id": {
                    "description": "Task identifier",
                    "type": "string",
                    "example": "abc123def456789"
                },
                "task_result": {
                    "description": "Query results (populated when status is SUCCESS)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/github_com_sudo-tiz_dns-tester-go_internal_models.DNSLookupResults"
                        }
                    ]
                },
                "task_status": {
                    "description": "Task status (PENDING, ACTIVE, SUCCESS, FAILURE)",
                    "type": "string",
                    "example": "SUCCESS"
                }
            }
        }
    },
    "tags": [
        {
            "description": "DNS lookup operations",
            "name": "DNS"
        },
        {
            "description": "Task management and status retrieval",
            "name": "Tasks"
        },
        {
            "description": "System health and metrics",
            "name": "System"
        }
    ]
}